### 2020-06-12 #2332 @whedon
[PRE REVIEW]: stella: Convolutional Neural Networks for Flare Identification in TESS
**Submitting author:** @afeinstein20 (<a href="http://orcid.org/0000-0002-9464-8101">Adina Feinstein</a>)
**Repository:** <a href="https://github.com/afeinstein20/stella" target ="_blank">https://github.com/afeinstein20/stella</a>
**Version:** v0.0.5
