### 2020-06-12 #2332 @afeinstein20
The proof looks good!
