### 2020-06-14 #2332 @arfon
:wave: @afeinstein20. Does this JOSS submission accompany a paper in a AAS journal?
