### 2020-06-16 #2347 @whedon
[REVIEW]: stella: Convolutional Neural Networks for Flare Identification in TESS
**Submitting author:** @afeinstein20 (<a href="http://orcid.org/0000-0002-9464-8101">Adina Feinstein</a>)
**Repository:** <a href="https://github.com/afeinstein20/stella" target ="_blank">https://github.com/afeinstein20/stella</a>
**Version:** v0.1.0

@pearsonkyle, @astrobel, & @ekaterinailin, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
