### 2020-06-16 #2347 @arfon
@pearsonkyle, @RuthAngus - One thing I forgot to mention when inviting you both to review is that this JOSS submission/paper is a joint publication with AAS Journals, i.e. this JOSS paper will be published together with another paper in one of the AAS journals. As part of this collaboration, AAS publishing makes a small donation to the running costs of JOSS.

If this is unacceptable to you, please let me know and I can look for alternative reviewers.

Some related links:

- Blog post describing the collaboration: https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
- Summary of JOSS financial model: https://joss.theoj.org/about#costs
