### 2020-06-16 #2347 @arfon
@RuthAngus - thanks for raising your potential conflict of interest here as a collaborator with @afeinstein20. As a result, I'd like to find an alternative reviewer - I'm sure it won't be too long before I ask you again 😉

:wave: @ekaterinailin & @astrobel - would either of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The package to be reviewed is _stella: Convolutional Neural Networks for Flare Identification in TESS_.
