### 2020-06-17 #2347 @arfon
✨ thanks @astrobel and @ekaterinailin! I think it would be great to have you both as reviewers here.

> Hi @arfon, thanks for the mention! I would be happy to help out and review the submission. Can I ask what the timeline is like for this review?

In the past we've asked for reviews to be completed in two weeks, but in our [_reduced service mode_](https://blog.joss.theoj.org/2020/05/reopening-joss) we're asking for reviews to be completed within six weeks instead.
