### 2020-06-17 #2347 @arfon
@pearsonkyle, @astrobel, @ekaterinailin - please carry out your review in this issue by updating the checklist above and giving feedback in this issue. The reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

Finally, as mentioned earlier in this thread, this JOSS submission/paper is a joint publication with AAS Journals, i.e. this JOSS paper will be published together with another paper in one of the AAS journals. As part of this collaboration, AAS publishing makes a small donation to the running costs of JOSS.

If this is unacceptable to you, please let me know and I can look for alternative reviewers.

Some related links:

Blog post describing the collaboration: https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
Summary of JOSS financial model: https://joss.theoj.org/about#costs
