### 2020-06-22 6377224
additional flare checks
