### 2020-06-25 afeinstein20/stella#6 @ekaterinailin
Tutorial: Creating A New CNN - clarification on data format needed
Hello again,

the Quickstart Tutorial went super smoothly :100: , so I am now at the second part **Creating A New CNN**. Here I got stuck at 1.1.

As I understood your example on the [docs](https://adina.feinste.in/stella/getting_started/tutorial.html) page, I need the path to a folder with light curves, and the path to a .csv file or similar that contains labeled flares.

So what I need to figure out is:

1. The light curves are supposed to be in .npy format. Should I download a bunch of TESS light cuves (as bulk or using lightkurve), and cast them in .npy with time, flux, and flux error indices?

2. What should the .csv file look like? Can I use Vizier table1 from [Günther+2020](https://ui.adsabs.harvard.edu/abs/2020AJ....159...60G/abstract)?

It would be awesome if you could drop a file or two as an example somewhere in the repo or elsewhere to get an idea. Any guidance in general is appreciated at this point :smiley_cat:

Cheers
Ekaterina
