### 2020-06-28 #2347 @pearsonkyle
Nice work @afeinstein20. Only a few things

## Software
- In the example, there was no way to inspect the model after it was loaded in `predict`. I'm curious about the number of inputs and size of the model being used. Someone not versed in keras/tensorflow might have difficulty finding these
- The dependencies are not clearly listed. There seems to be a disconnect between `requirements.txt` and `setup.py`
- The example works well and was straight forward, nice job.
- The automated tests could use some work - this is to make sure any new changes don't break the core functionality

## Paper
- no mention of previous work done in detecting flares. Specifically, Kepler which had a similar data set and was around before TESS. Are you able to add a figure?

## General
Maybe these will be answered in the science paper and so I apologize ahead of time, I don't think we have access to that.
- You mentioned flare events have a particular time scale, what is the min/max flare duration the CNN is sensitive to? Would something like a cosmic ray cause a false positive? What about an exoplanet transit? Star spot?
- How much preprocessing is required for the best results? e.g. 3-sigma clip time series, detrend and then run?
- If you don't detrend, how did you account for this in the training data? since stars are variable
- How well does the code handle gaps in the data?
- How does it handle data sets at different cadences? Or is the code designed for TESS 2 minute observations?
