### 2020-07-02 #2347 @ekaterinailin
Thank you for making the effort of sharing your software in a usable way, @afeinstein It's very much appreciated! Installation was easy, and the quickstart tutorial is neat and worked smoothly. :100:

I will be very favorable of acceptance if my major concerns are addressed:

Sorted by importance:

- Major concerns:

  - The test suite needs come actual tests, both unit and integration tests.
  - I could not complete the Build your own CNN tutorial based on the instructions in the docs (see Issue in the repo [here](https://github.com/afeinstein20/stella/issues/6)).
  - requirements.txt is incomplete compared to travis.yml and setup.py

-  Minor issues:
   - Can you add the usage of the FitFlares class to the docs if you consider them core functionalities of your software?
   - In the paper, the phrase "rotation periods, large spot modulation, and increased flare rates" should cite at least a review article or similar.
   - In the docs, a statement about how I can 1) contribute to the software 2) report issues or problems with the software 3) seek support in an easily accessible place would be useful.
   - In the paper, the paragraph on previous flare finding techniques could be expanded with references that partially appear in the preprint of Feinstein et al. (2020). Especially those that use machine learning algorithms like Vida and Roettenbacher (2018). See their Github [here](https://github.com/vidakris/flatwrm/).

- Nice to have:
  - The API in the docs would benefit from an index pointing to each of the five classes in your software.

(I left the boxes that relate to my major concerns unchecked, is that how I should do it, @arfon ? Or should I check them anyways to indicate that I addressed them?).
