### 2020-07-02 #2347 @arfon
> (I left the boxes that relate to my major concerns unchecked, is that how I should do it, @arfon ? Or should I check them anyways to indicate that I addressed them?).

No, you've done it correctly. As the author addresses your feedback, we'd ask you to update the checklist.
