### 2020-07-08 #2347 @astrobel
Hi @afeinstein20! Really like this software, and I think your paper & documentation are great. In particular, I think the documentation is really friendly to beginners, with regards to both the science and the code. I have a couple of major notes below, and then just some really minor stuff which I think would help with clarity and consistency, especially in the docs.

## Major notes:

- **Installation instructions**: I had no trouble installing the software or dependencies, however it seems like `requirements.txt` is incomplete. I would either remove that file or update it to match the Travis list.
- **Automated tests**: The test suite needs substantial expansion to cover the software's full functionality.
- **Community guidelines**: The package needs clear community guidelines for contribution, to go alongside the thorough code of conduct. These could both be linked in the README.
- **State of the field**: You've got some really good info on previous methods of flare detection in the second paragraph of the paper, but I would suggest including some citations here to other work. I would also be interested to know if any of these previous methods are also open-source and freely available!

## Minor notes:

### General

- In the first sentence of the third paragraph in the paper, the bracketed abbreviation should be CNNs to reflect its plural use in the sentence.
- When you refer to classes in Lightkurve and Astropy, it would be good to link to their APIs.

### Quickstart Tutorial
- The standard for Lightkurve tutorials is `import lightkurve as lk`, so it might reduce confusion if you change the name of your light curve variable.
- A couple of deprecation warnings in `[7]`: `search_lightcurvefile` should be replaced with `search_lightcurve`. Also worth keeping in mind that the `.PDCSAP_FLUX` method is deprecated, but I can't find a good replacement for now without significant changes so I reckon just keep an eye out on this one.
- This is really tiny, but in `[9]` you should use the `single_pred` variable for the scatter plot colour scale, which will make it a nice parallel with `average_pred` below.
- To clarify, when you use ensembling, you're using a median in the code and not an average? If a median is indeed what's needed, this should be reflected in the text.

### Creating a CNN
- As the catalogue data isn't available yet, I haven't been able to work through this section of the tutorial.
- I'm a little confused by the paragraph about light curve pre-processing, and I can't pinpoint the variables with default settings which you mention. It would be great to link to the API and the exact function/arguments you're talking about.
- There's some overlap between Section 1.4 here and the quickstart tutorial. The material is useful to include in both, but I think it could be helpful to signal somewhere in the CNN tutorial that they cover the same material.
