### 2020-07-13 f7fa41f
new class for downloading the catalog and light curves
