### 2020-07-13 afeinstein20/stella#6 @afeinstein20
Hi @ekaterinailin Thanks for pointing this out!! I have added a class called "DownloadSets" which will now download the flare catalog from Günther+2020 from Vizier and download/reformat the light curves for the CNN training sets. The tutorial notebook has been updated to reflect this as well :)

Please close the issue if you find it works!
