+++ requirements.txt
@@ -1,3 +1,20 @@
 nbsphinx
 ipykernel
 pylab
+tensorflow
+astroquery
+astropy
+tqdm
+cython
+codecov
+coverage
+pytest
+pytest-cov
+lxml
+pybind11
+tornado
+matplotlib
+lightkurve
+numpy
+sklearn
+more-itertools
