### 2020-07-13 e868fec
FOR LIGHTKURVE
