### 2020-07-13 1904a18
version conflict
lightkurve vs. tensorflow
