+++ setup.py
@@ -46,7 +46,7 @@ setup(
         'astroquery', 'sklearn', 
         'setuptools>=41.0.0', 'more-itertools',
         'matplotlib', 'numpy', 'scipy>=1.5.0',
-        'tensorflow===2.1.0', 'lightkurve'],
+        'tensorflow'],#, 'lightkurve'],
     classifiers=[
         'Development Status :: 3 - Alpha',
         'Intended Audience :: Science/Research',
