### 2020-07-13 9a2f573
which lightkurve works
