+++ stella/tests/test_neural_network.py
@@ -1,5 +1,1 @@
-import numpy as np
-
-
-def test_anything():
-    assert(1==1)
+from stella import ConvNN
