+++ paper.bib
@@ -36,7 +36,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{Guenther:2020,
+@ARTICLE{guenther:2020,
        author = {{G{\"u}nther}, Maximilian N. and {Zhan}, Zhuchang and {Seager}, Sara and
          {Rimmer}, Paul B. and {Ranjan}, Sukrit and {Stassun}, Keivan G. and
          {Oelkers}, Ryan J. and {Daylan}, Tansu and {Newton}, Elisabeth and
