+++ paper.md
@@ -42,14 +42,15 @@ Flare rates and energies can yield consequences for the early stages of planet f
 Models have demonstrated that the introduction of superflares ($> 5\%$ flux increase) are able to irreparably alter the chemistry of an atmosphere [@venot:2016] and expedite atmospheric photoevaporation [@lammer:2007]. 
 Thus, understanding flare rates and energies at young ages provides crucial keys for understanding the exoplanet population we see today.
 
-Previous methods of flare detection have relied on detrending a light curve and using outlier detection heuristics for identifying flare events [@davenport:2016, @vida18, @allesfitter].
+Previous methods of flare detection have relied on detrending a light curve and using outlier detection heuristics for identifying flare events [@Davenport:2016],[@allesfitter]. 
+More complex methods, such as a RANdom SAmple Consensus (RANSAC) algorithm has been tested as well [@vida18]. RANSAC algorithms identfy and subtract inliers (the underlying light curve) before searching for outliers above a given detection threshold.
 Low-amplitude flares can easily be removed with aggressive detrending techniques (e.g. using a small window-length to remove spot modulation). 
 Additionally, low energy flares likely fall below the outlier threshold, biasing the overall flare sample towards higher energy flares.
 As flares exhibit similar temporal evolution (a sharp rise followed by an exponential decay, with the exception of complex flare groups), machine learning algorithms may prove suitable for identifying such features without light curve detrending.
 
 `stella` is an open-source Python package for identifying flares in the *Transiting Exoplanet Survey Satellite* (*TESS*) two-minute data with convolutional neural networks (CNNs).
 Users have the option to use the models created in @Feinstein:2020 or build their own cutomized networks.
-The training, validation, and test sets for our CNNs use the flare catalog presented in @guenther:2019. These light curves are publicly available through the Mikulski Archive for Space Telescopes and can be downloaded through the \texttt{lightkurve} package [@lightkurve]; they are not included in \texttt{stella}.
+The training, validation, and test sets for our CNNs use the flare catalog presented in @guenther:2020. These light curves are publicly available through the Mikulski Archive for Space Telescopes and can be downloaded through the \texttt{lightkurve} package [@lightkurve]; they are not included in \texttt{stella}.
 It takes approximately twenty minutes to train a \texttt{stella} model from scratch and $<1$ minute to predict flares on a single sector light curve.
 The package also allows users to measure rotation periods and fit flares to extract underlying flare parameters. Further documentation and tutorials can be found at \url{adina.feinste.in/stella}.
 
