+++ paper.bib
@@ -186,3 +186,82 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2019A&A...622A.133I},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@ARTICLE{borucki:2010,
+       author = {{Borucki}, William J. and {Koch}, David and {Basri}, Gibor and
+         {Batalha}, Natalie and {Brown}, Timothy and {Caldwell}, Douglas and
+         {Caldwell}, John and {Christensen-Dalsgaard}, J{\o}rgen and
+         {Cochran}, William D. and {DeVore}, Edna and {Dunham}, Edward W. and
+         {Dupree}, Andrea K. and {Gautier}, Thomas N. and {Geary}, John C. and
+         {Gilliland}, Ronald and {Gould}, Alan and {Howell}, Steve B. and
+         {Jenkins}, Jon M. and {Kondo}, Yoji and {Latham}, David W. and
+         {Marcy}, Geoffrey W. and {Meibom}, S{\o}ren and {Kjeldsen}, Hans and
+         {Lissauer}, Jack J. and {Monet}, David G. and {Morrison}, David and
+         {Sasselov}, Dimitar and {Tarter}, Jill and {Boss}, Alan and
+         {Brownlee}, Don and {Owen}, Toby and {Buzasi}, Derek and
+         {Charbonneau}, David and {Doyle}, Laurance and {Fortney}, Jonathan and
+         {Ford}, Eric B. and {Holman}, Matthew J. and {Seager}, Sara and
+         {Steffen}, Jason H. and {Welsh}, William F. and {Rowe}, Jason and
+         {Anderson}, Howard and {Buchhave}, Lars and {Ciardi}, David and
+         {Walkowicz}, Lucianne and {Sherry}, William and {Horch}, Elliott and
+         {Isaacson}, Howard and {Everett}, Mark E. and {Fischer}, Debra and
+         {Torres}, Guillermo and {Johnson}, John Asher and {Endl}, Michael and
+         {MacQueen}, Phillip and {Bryson}, Stephen T. and {Dotson}, Jessie and
+         {Haas}, Michael and {Kolodziejczak}, Jeffrey and {Van Cleve}, Jeffrey and
+         {Chandrasekaran}, Hema and {Twicken}, Joseph D. and
+         {Quintana}, Elisa V. and {Clarke}, Bruce D. and {Allen}, Christopher and
+         {Li}, Jie and {Wu}, Haley and {Tenenbaum}, Peter and
+         {Verner}, Ekaterina and {Bruhweiler}, Frederick and {Barnes}, Jason and
+         {Prsa}, Andrej},
+        title = "{Kepler Planet-Detection Mission: Introduction and First Results}",
+      journal = {Science},
+     keywords = {ASTRONOMY},
+         year = 2010,
+        month = feb,
+       volume = {327},
+       number = {5968},
+        pages = {977},
+          doi = {10.1126/science.1185402},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2010Sci...327..977B},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@INPROCEEDINGS{ricker:2014,
+       author = {{Ricker}, George R. and {Winn}, Joshua N. and {Vanderspek}, Roland and
+         {Latham}, David W. and {Bakos}, G{\'a}sp{\'a}r. {\'A}. and
+         {Bean}, Jacob L. and {Berta-Thompson}, Zachory K. and
+         {Brown}, Timothy M. and {Buchhave}, Lars and {Butler}, Nathaniel R. and
[…]
