### 2020-07-13 #2347 @afeinstein20
Hi @pearsonkyle Thanks for your review! I'm addressing your comments below:

**Software**
- I've included the loaded model as an attribute in the `stella.ConvNN.predict()` function. I've also added a demonstration on how to access the model summary, which includes information on the CNN layers, sizes, input shapes, etc. in the [Quickstart Tutorial](http://adina.feinste.in/stella/getting_started/shortest_demo.html). If there are any additional demonstrations you think would be useful, I'm happy to include them here as well.
- I have fixed the disconnect between `requirements.txt` and `setup.py`.
- I have included some better tests now.

**Paper**
- I have included additional citations to open-source codes for flare detections and wrote a bit more about some of those specific methods.

**General**
The paper can be found [here](https://arxiv.org/abs/2005.07710).
- Cosmic rays are removed with quality flags, so I'm not sure how the CNN would react if it was included. It generally does not mark 1-cadence outliers, which suggests it is doing a good job at identifying the typical flare-shape. It also did not find any transits (that's my next project). As for star spots, it does a good job at not identifying spot modulation as flares with the exception of very rapid rotators (< 0.5 days) because those tend to look "sharper" and more flare-like at the peak.
- No preprocessing is required beyond using the _TESS_ provided quality flags.
- The training data was also not detrended in any way. Because the training set looked at every star in Sectors 1 & 2 at 2-minute cadence, there were variable stars included in that set, which provided a good set for the CNN to learn that spots != flares and therefore distinguish between the two.
- We ignore gaps and the 100 cadences closest to the gaps. In Figure 12 of the paper, you can see this in the assigned colors (the ignored points are light green near the gaps).
- This was specifically designed for _TESS_ 2-minute observations.
