+++ stella/tests/test_download.py
@@ -1,5 +1,5 @@
 from stella.download_nn_set import *
-from stella.preprocessing_flares.py import *
+from stella.preprocessing_flares import *
 from numpy.testing import assert_almost_equal
 
 download = DownloadSets(fn_dir='.')
