### 2020-07-14 #2347 @afeinstein20
Hi @ekaterinailin Thanks for your review! I'm addressing your comments below:

**Major Concerns**
- I've incorporated better tests now to ensure the core functionality doesn't break.
- I have created a new class `stella.DownloadSet` which allows the user to download the Günther+2020 flare catalog from Vizier and the light curves for the CNN training/validation/test sets through `lightkurve`. A demo of how this class works is presented in the [Creating a New CNN](http://adina.feinste.in/stella/getting_started/tutorial.html) notebook.
- I have fixed both `requirements.txt` and `.travis.yml` to be consistent.

**Minor Issues**
- Thank you for pointing out that there is no demo for `stella.FitFlares`! I have gone and incorporated that in the [Other Fun Things](http://adina.feinste.in/stella/getting_started/other_features.html) notebook.
- I have added appropriate citations as requested.
- There is a short paragraph describing issues/contributions on the [main documentation page](http://adina.feinste.in/stella/index.html). Is this sufficient, or were you envisioning something more?
- I have added appropriate citations to other open-source flare detection methods, including Appaloosa, AllesFitter, and Flatw'rm.
