### 2020-07-14 afeinstein20/stella#7 @afeinstein20
Updated lightkurve in setup.py
Need to wait for tensorflow >= 2.3.0 to update lightkurve. The following changes need to be made when this update is implemented:

- `search_lightcurvefile `should be replaced with `search_lightcurve`.
- `.PDCSAP_FLUX` method is deprecated, but I can't find a good replacement for now without significant changes so I reckon just keep an eye out on this one.
