### 2020-07-14 #2347 @afeinstein20
Hi @astrobel Thanks for your review! I'm addressing your comments below:

### Major Comments
- I have addressed the incompleteness and inconsistencies between `requirements.txt`, `.travis.yml`, and `setup.py` files.
- I have included better tests now.
- There is a short section for contributions/reporting issues on the [documentation](http://adina.feinste.in/stella/) page, and I have copied this over to the README so it is more obvious to users. Thanks for this suggestion!
- I have included references and mentioned additional open-source flare detection software, including Appaloosa, Flatw'rm, and AllesFitter.

### Minor Notes

**Quickstart Tutorial**
- Oop, I didn't know that was standard... I will change the variable name. Thanks!
- Thanks for pointing out these issues with `lightkurve`! Unfortunately, there is currently a dependency conflict for scipy between lightkurve (requires v >= 1.5.0) and tensorflow (requires v == 1.4.1). This is a [noted issue](https://github.com/tensorflow/tensorflow/pull/40789) and will be fixed in the next version update for tensorflow. Because tensorflow is the backbone of stella and the dependencies need to work for this package, I have included `lightcurve<=1.9.0` in the `setup.py`. I have copied over your notes into a [stella issue](https://github.com/afeinstein20/stella/issues/7) so that when tensorflow is updated, I can make these changes.
- I agree comparing the `single_pred` to `average_pred` would make a nice addition and have included it in the [Quickstart tutorial](http://adina.feinste.in/stella/getting_started/shortest_demo.html).

**Creating a CNN**
- I have created a new class `stella.DownloadSets()` which will query the flare catalog from Günther+2020 through Vizier and the appropriate light curves for the training/validation/test sets through `lightkurve`. A demonstration on how this class works is presented in [associated demo notebook](http://adina.feinste.in/stella/getting_started/tutorial.html).
- I have included a link to the API and added some additional text, with the hopes this clarifies the pre-processing variables.
