### 2020-07-15 #2347 @afeinstein20
> > Hi @arfon ! I've sent responses to each of the referees :) Is there anything else I need to/can do right now?
>
> No, that's perfect 👍 - let's wait for their responses now.

Sounds good! I also wanted to let you know I just received the 2nd referee report for the associated AAS paper. It's quite short and I hope to resubmit soon. Hopefully the timelines for this submission and that one will line up okay!!
