### 2020-07-16 #2347 @astrobel
Thanks for your thorough response, @afeinstein20! Everything is looking good, and I've now checked off the remaining boxes on the reviewer checklist. I'm also really excited to see `stella.DownloadSets()` in there, and I'm hoping to play around with it soon!

Just one thing, I don't think you've pushed the changes to the README with the community guidelines updated?
