### 2020-07-16 b0c084a
update paper branch
