### 2020-07-19 afeinstein20/stella#9 @ekaterinailin
ff.flare_table comes out with fewer flares than the tutorial indicates
Hi @afeinstein20 !

I tried to use the `FitFlares` feature, but got stuck at the point where I wanted to show the table with the fitted flares.

This is the code I ran (trying to reproduce the Fitting Flares tutorial):

```
import stella

# Get the light curve

from lightkurve.search import search_lightcurvefile

lk = search_lightcurvefile(target='tic62124646', mission='TESS')
lk = lk.download().PDCSAP_FLUX

# get the model

cnn = stella.ConvNN(output_dir=OUT_DIR)

MODEL_DIR = 'run01/'
MODEL = [os.path.join(MODEL_DIR,i) for i in
          os.listdir(MODEL_DIR) if i.endswith('.h5')][0]

# predict flares from model

cnn.predict(modelname=MODEL,
            times=lk.time,
            fluxes=lk.flux,
            errs=lk.flux_err)

# fit flares

ff = stella.FitFlares(id=[lk.targetid],
                      time=[lk.time],
                      flux=[lk.flux],
                      flux_err=[lk.flux_err],
                      predictions=[cnn.predictions[0]])

# get flare table

ff.identify_flare_peaks(threshold=0.5)

ff.flare_table
```

The output is a single flares

Target_ID | tpeak | amp | ed_s | rise | fall | prob
-- | -- | -- | -- | -- | -- | --
62124646.0 | 1662.6966771763764 | 1.0031577638536382 | 2.5355761288160976 | 0.0028550217295811296 | 0.008296178315620328 | 0.976371169090271

which is not what I expected. Based on the `cnn.predictions` I expected a number of flares similar to what your tutorial shows (16?).

I tried several models from the quickstart collection, but they either return 1 or zero flares in the `ff.flare_table`.

Do you have an idea what I am doing wrong here?

Cheers :)
Ekaterina
