### 2020-07-19 #2347 @ekaterinailin
Hi @afeinstein,

thanks for addressing the comments! The software performs very well now, and the paper is good to go. However, I still think that the tests require some more attention, so I have not ticked this last box yet.

Responding to each of your replies one by one:

**Major Issues**

- I've incorporated better tests now to ensure the core functionality doesn't break.

The tests that you added to the master branch look good but they do not run through in Travis. Your [last build on the master branch](https://travis-ci.org/github/afeinstein20/stella/builds/708191753) errorred with this message:

```
$ pytest stella/tests/* --cov=stella/

rootdir: /home/travis/build/[secure]/stella, inifile: setup.cfg

plugins: cov-2.10.0

collected 0 items

============================ no tests ran in 0.01s =============================

ERROR: not found: /home/travis/build/[secure]/stella/stella/tests/ensemble_s0005_i0350_b0.73.h5

(no name '/home/travis/build/[secure]/stella/stella/tests/ensemble_s0005_i0350_b0.73.h5' in any of [])
```

Once this is fixed, I also suggest that you add tests of the most important functions in the modules `metrics.py`, `utils.py`, and `mark_flares.py`.

_Side note: A convention that I find useful when writing tests: Each module `name.py` should correspond to a test module `test_name.py` and only test the classes and functions that are defined in `name.py`. This makes it easier to navigate the tests in the long run._

- I have created a new class stella.DownloadSet which allows the user to download the Günther+2020 flare catalog from Vizier and the light curves for the CNN training/validation/test sets through lightkurve. A demo of how this class works is presented in the Creating a New CNN notebook.

Great additon :100: , now the tutorial runs smoothly. Looking forward to using my own flare tables! Closed the issue.

- I have fixed both requirements.txt and .travis.yml to be consistent.

:+1:

**Minor Issues**

- Thank you for pointing out that there is no demo for stella.FitFlares! I have gone and incorporated that in the Other Fun Things notebook.

I tried the tutorial but got stuck at one point. I opened an [issue](https://github.com/afeinstein20/stella/issues/9) on that, but I suspect that the error is on my side here.

- There is a short paragraph describing issues/contributions on the main documentation page. Is this sufficient, or were you envisioning something more?

Ah, found it! I think that directing people to Github should be fine in general. In my experience, however, users find opening an issue quite a barrier for interaction, which is why I would recommend adding an email address to the paragraph. I even think you could make it more accessible if you dedicated a category on the navigation bar on the left to it.

- I have added appropriate citations as requested.

:+1:

- I have added appropriate citations to other open-source flare detection methods, including Appaloosa, AllesFitter, and Flatw'rm.

:+1:
