### 2020-07-21 afeinstein20/stella#9 @afeinstein20
Hi @ekaterinailin ! Weird. Which version of stella are you using? If you're using the version that was `pip install stella` then I would believe there would be differences between that and the version now on GitHub.

I will be updating the PyPI version once the referee report is good to go and the associated AJ paper is accepted!
