### 2020-07-21 #2347 @afeinstein20
> Thanks for your thorough response, @afeinstein20! Everything is looking good, and I've now checked off the remaining boxes on the reviewer checklist. I'm also really excited to see `stella.DownloadSets()` in there, and I'm hoping to play around with it soon!
>
> Just one thing, I don't think you've pushed the changes to the README with the community guidelines noted?

Ah yes thanks for pointing this out! This is what I get for trying to work in different branches. The README now has the guidelines.
