+++ .travis.yml
@@ -15,6 +15,9 @@ addons:
 
 cache: pip
 
+file:
+  - ensemble_s0005_i0350_b0.73.h5
+
 before_install:
   - sudo apt-get install python3-pip
   - sudo pip3 install requests
