+++ .travis.yml
@@ -16,7 +16,7 @@ addons:
 cache: pip
 
 file:
-  - "ensemble_s0005_i0350_b0.73.h5"
+  - "stella/tests/ensemble_s0005_i0350_b0.73.h5"
 
 before_install:
   - sudo apt-get install python3-pip
