+++ .travis.yml
@@ -1,9 +1,7 @@
-dist: xenial
-sudo: true
+dist: trusty
 
 language: python
-python: 
-  - "3.7"
+python: 3.7
 
 matrix:
   include:
