+++ .travis.yml
@@ -1,7 +1,7 @@
 dist: trusty
 
 language: python
-python: 3.7
+python: 3.6
 
 matrix:
   include:
