+++ .travis.yml
@@ -1,7 +1,7 @@
-dist: trusty
+dist: xenial
 
 language: python
-python: 3.6
+python: 3.7
 
 matrix:
   include:
