+++ .travis.yml
@@ -10,8 +10,6 @@ matrix:
 addons:
   apt_packages:
     - pandoc
-  file:
-    - "stella/tests/ensemble_s0005_i0350_b0.73.h5"
 
 cache: pip
 
