+++ stella.egg-info/requires.txt
@@ -5,4 +5,7 @@ sklearn
 setuptools>=41.0.0
 more-itertools
 matplotlib
-tensorflow>=2.0.0
+numpy
+scipy==1.4.1
+tensorflow>2.0.0
+lightkurve==1.9.0
