### 2020-07-22 afeinstein20/stella#9 @ekaterinailin
I think I tried both the pip version and the cloned repo. I now reproduced the same result with the most recent master branch version.
