### 2020-07-22 afeinstein20/stella#9 @afeinstein20
Okay great! Yeah I think there was a flub in me working in different branches, but now there's consistency between them. Please close the issue if you're satisfied with the fix!
