+++ stella/tests/test_download.py
@@ -3,6 +3,7 @@ from stella.preprocessing_flares import *
 from stella.neural_network import *
 from numpy.testing import assert_almost_equal
 
+
 download = DownloadSets(fn_dir='.')
 download.download_catalog()
 download.download_lightcurves()
