+++ stella/tests/test_download.py
@@ -8,8 +8,8 @@ download.flare_table = download.flare_table[0:20]
 
 def test_catalog_retrieval():
     assert(download.flare_table['TIC'][0] == 2760232)
-    assert_almost_equal(download.flare_table['tpeak'][100], 2458379.9, decimal=1)
-    assert(download.flare_table['Flare'][4000] == 1)
+    assert_almost_equal(download.flare_table['tpeak'][10], 2458368.8, decimal=1)
+    assert(download.flare_table['Flare'][9] == 3)
 
 def test_lightcurves():
     download.download_lightcurves()
