+++ stella/tests/test_download.py
@@ -5,17 +5,15 @@ from numpy.testing import assert_almost_equal
 download = DownloadSets(fn_dir='.')
 download.download_catalog()
 download.flare_table = download.flare_table[0:20]
+download.download_lightcurves()
+pre = FlareDataSet(downloadSet=download)
 
 def test_catalog_retrieval():
     assert(download.flare_table['TIC'][0] == 2760232)
     assert_almost_equal(download.flare_table['tpeak'][10], 2458368.8, decimal=1)
     assert(download.flare_table['Flare'][9] == 3)
 
-def test_lightcurves():
-    download.download_lightcurves()
-
 def test_processing():
-    pre = FlareDataSet(downloadSet=download)
     assert_almost_equal(pre.frac_balance, 0.7, decimal=1)
     assert(pre.train_data.shape == (48, 200, 1))
     assert(pre.val_data.shape == (6, 200, 1))
