+++ stella/tests/test_download.py
@@ -1,5 +1,6 @@
 from stella.download_nn_set import *
 from stella.preprocessing_flares import *
+from stella.neural_network import *
 from numpy.testing import assert_almost_equal
 
 download = DownloadSets(fn_dir='.')
@@ -23,3 +24,13 @@ def test_tensorflow():
     import tensorflow
     assert(tensorflow.__version__ == '2.1.0')
 
+def test_build_model():
+    cnn = ConvNN(output_dir='.',
+                 ds=pre)
+    cnn.train_models(epochs=10)
+
+    assert(cnn.loss == 'binary_crossentropy')
+    assert(cnn.optimizer == 'adam')
+    assert(cnn.training_ids[10] == 2760232.0)
+    assert(cnn.frac_balance == 0.73)
+    assert(len(cnn.val_pred_table) == 6)
