+++ stella/tests/test_preprocessing.py
@@ -1,11 +1,0 @@
-from stella.preprocessing_flares import *
-from numpy.testing import assert_almost_equal
-
-pre = FlareDataSet(fn_dir = '.',
-                   catalog = 'Guenther_2020_flare_catalog.txt')
-
-def test_processing():
-    assert_almost_equal(pre.frac_balance, 0.7, decimal=1)
-    assert(pre.train_data.shape == (48, 200, 1))
-    assert(pre.val_data.shape == (6, 200, 1))
-    assert(pre.test_data.shape == (7, 200, 1))
