+++ stella/tests/test_download.py
@@ -19,9 +19,9 @@ def test_light_curves():
 
 def test_processing():
     assert_almost_equal(pre.frac_balance, 0.7, decimal=1)
-    assert(pre.train_data.shape == (48, 200, 1))
-    assert(pre.val_data.shape == (6, 200, 1))
-    assert(pre.test_data.shape == (7, 200, 1))
+    assert(pre.train_data.shape == (62, 200, 1))
+    assert(pre.val_data.shape == (8, 200, 1))
+    assert(pre.test_data.shape == (8, 200, 1))
 
 def test_tensorflow():
     import tensorflow
