+++ stella/tests/test_processing.py
@@ -6,6 +6,6 @@ pre = FlareDataSet(fn_dir='.',
 
 def test_processing():
     assert_almost_equal(pre.frac_balance, 0.7, decimal=1)
-    assert(pre.train_data.shape == (67, 200, 1))
+    assert(pre.train_data.shape == (62, 200, 1))
     assert(pre.val_data.shape == (8, 200, 1))
     assert(pre.test_data.shape == (8, 200, 1))
