+++ MANIFEST.in
@@ -1,3 +1,2 @@
 include README.md 
 include LICENSE 
-include stella/ensemble_s0005_i0350_b0.73.h5
