### 2020-07-29 #2347 @afeinstein20
Hi @ekaterinailin ! I've addressed your comments below:

 **Major Issues**
> The tests that you added to the master branch look good but they do not run through in Travis. Your [last build on the master branch](https://travis-ci.org/github/afeinstein20/stella/builds/708191753) errorred with this message:

- Travis has been *finally* fixed!

> Once this is fixed, I also suggest that you add tests of the most important functions in the modules `metrics.py`, `utils.py`, and `mark_flares.py`.

- Many of the functionalities in `utils.py` are tested in `test_neural_network.py`, specifically the functions `break_rest`, `do_the_shuffle`, `split_data` as these are all crucial steps for preprocessing the data to feed into the neural network. Additionally, `utils.flare_lightcurve` is tested in the new script `test_mark_flares.py`.

> _Side note: A convention that I find useful when writing tests: Each module `name.py` should correspond to a test module `test_name.py` and only test the classes and functions that are defined in `name.py`. This makes it easier to navigate the tests in the long run.

- I agree this is the typical convention! However Travis seems to have trouble saving/loading/locating Tensorflow models, so to mitigate this issue, I've tested `stella.DownloadSets`, `stella.FlareDataSet`, `stella.ConvNN`, and `stella.ModelMetrics` in the same test python script. I've split up the others where I could.

**Minor Issues**

> Ah, found it! I think that directing people to Github should be fine in general. In my experience, however, users find opening an issue quite a barrier for interaction, which is why I would recommend adding an email address to the paragraph. I even think you could make it more accessible if you dedicated a category on the navigation bar on the left to it.

- I generally disagree with this. I've hyperlinked the GitHub Issues page in the documentation, so this should make it easier for users. Submitting a GitHub issue rather than emailing me directly is beneficial for users and me in several ways. For one, it allows other users to see and comment if they are having similar issues. Closed issues are also visible, which is a useful resource for new users to the code. Second, it keeps a clean running list of things I need to fix, rather than having many different emails and potentially losing track of things. (Also, I've dealt with a lot of repeat issue emails from eleanor, which could have been mitigated if users had submitted issues for other users to see!)
