### 2020-07-31 #2347 @ekaterinailin
Hi @afeinstein20 !

Thanks for addressing my extra comments. Test are good to go, so I checked the last box :+1: . May this software be useful for many flare researchers to come :crossed_fingers: .

**re: Github issues vs. email** - I think you're making a fair point. Eventually, it's up to the maintainer of the software to choose what works best for them.
