### 2020-07-31 #2347 @arfon
👋 @afeinstein20 - looks like this submission is pretty much ready to accept and publish on the JOSS side of things.

How is the AAS paper coming along? Have you received any notifications from them about what the DOI will be for that paper? As this is a joint AAS/JOSS paper, we'll need to do a little coordination with AAS publishing before we publish here to ensure that the JOSS paper correctly cross-links to the AAS paper.
