### 2020-07-31 afeinstein20/stella#9 @ekaterinailin
I got a fresh installation from the master branch now, but unfortunately to no avail. I get a bunch of outputs tothe prompt, not sure if related to the actual issue, but perhaps worth posting anyways. Some I expected from the tutorial, others not:

```
/home/ekaterina/Documents/000_envs/stella/lib/python3.6/site-packages/lightkurve/lightcurvefile.py:47: LightkurveWarning: `LightCurveFile.header` is deprecated, please use `LightCurveFile.get_header()` instead.
  LightkurveWarning)

WARNING: No stella.DataSet object passed in.
Can only use stella.ConvNN.predict().

100%|██████████| 1/1 [00:01<00:00,  1.13s/it]
Finding & Fitting Flares:   0%|          | 0/1 [00:00<?, ?it/s]/home/ekaterina/Documents/000_envs/stella/lib/python3.6/site-packages/stella-0.1.0rc1-py3.6.egg/stella/mark_flares.py:92: VisibleDeprecationWarning: Creating an ndarray from ragged nested sequences (which is a list-or-tuple of lists-or-tuples-or ndarrays with different lengths or shapes) is deprecated. If you meant to do this, you must specify 'dtype=object' when creating the ndarray
  return np.array(results)
/home/ekaterina/Documents/000_envs/stella/lib/python3.6/site-packages/numpy/lib/nanfunctions.py:1665: RuntimeWarning: Degrees of freedom <= 0 for slice.
  keepdims=keepdims)
/home/ekaterina/Documents/000_envs/stella/lib/python3.6/site-packages/numpy/lib/nanfunctions.py:1114: RuntimeWarning: All-NaN slice encountered
  overwrite_input=overwrite_input)
Finding & Fitting Flares: 100%|██████████| 1/1 [00:00<00:00,  5.22it/s]
```
