### 2020-07-31 #2347 @arfon
All we'll need to do then is get the DOI from AAS before we accept and publish here.
