+++ paper.bib
@@ -36,28 +36,34 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{guenther:2019,
+@ARTICLE{guenther:2020,
        author = {{G{\"u}nther}, Maximilian N. and {Zhan}, Zhuchang and {Seager}, Sara and
          {Rimmer}, Paul B. and {Ranjan}, Sukrit and {Stassun}, Keivan G. and
          {Oelkers}, Ryan J. and {Daylan}, Tansu and {Newton}, Elisabeth and
-         {Gillen}, Edward and {Rappaport}, Saul and {Ricker}, George R. and
+         {Kristiansen}, Martti H. and {Olah}, Katalin and {Gillen}, Edward and
+         {Rappaport}, Saul and {Ricker}, George R. and {Vanderspek}, Roland K. and
          {Latham}, David W. and {Winn}, Joshua N. and {Jenkins}, Jon M. and
          {Glidden}, Ana and {Fausnaugh}, Michael and {Levine}, Alan M. and
          {Dittmann}, Jason A. and {Quinn}, Samuel N. and
          {Krishnamurthy}, Akshata and {Ting}, Eric B.},
-        title = "{Stellar Flares from the First Tess Data Release: Exploring a New Sample of M-dwarfs}",
-      journal = {arXiv e-prints},
+        title = "{Stellar Flares from the First TESS Data Release: Exploring a New Sample of M Dwarfs}",
+      journal = {\aj},
      keywords = {Astrophysics - Earth and Planetary Astrophysics},
-         year = "2019",
-        month = "Jan",
-          eid = {arXiv:1901.00443},
-        pages = {arXiv:1901.00443},
+         year = 2020,
+        month = feb,
+       volume = {159},
+       number = {2},
+          eid = {60},
+        pages = {60},
+          doi = {10.3847/1538-3881/ab5d3a},
 archivePrefix = {arXiv},
        eprint = {1901.00443},
  primaryClass = {astro-ph.EP},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2019arXiv190100443G},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2020AJ....159...60G},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+
 
 @ARTICLE{Feinstein:2020,
        author = {{Feinstein}, Adina D. and {Montet}, Benjamin T. and {Ansdell}, Megan and
@@ -74,6 +80,188 @@ archivePrefix = {arXiv},
        eprint = {2005.07710},
  primaryClass = {astro-ph.SR},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2020arXiv200507710F},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@MISC{lightkurve,
+   author = {{Lightkurve Collaboration} and {Cardoso}, J.~V.~d.~M. and
+             {Hedges}, C. and {Gully-Santiago}, M. and {Saunders}, N. and
+             {Cody}, A.~M. and {Barclay}, T. and {Hall}, O. and
+             {Sagear}, S. and {Turtelboom}, E. and {Zhang}, J. and
+             {Tzanidakis}, A. and {Mighell}, K. and {Coughlin}, J. and
+             {Bell}, K. and {Berta-Thompson}, Z. and {Williams}, P. and
+             {Dotson}, J. and {Barentsen}, G.},
+    title = "{Lightkurve: Kepler and TESS time series analysis in Python}",
+ keywords = {Software, NASA},
+howpublished = {Astrophysics Source Code Library},
+     year = 2018,
[…]
