+++ stella.egg-info/requires.txt
@@ -7,5 +7,5 @@ more-itertools
 matplotlib
 numpy
 scipy==1.4.1
-tensorflow>2.0.0
+tensorflow==2.1.0
 lightkurve==1.9.0
