### 2020-08-03 #2347 @afeinstein20
Hi @arfon !

The AAS paper has been officially accepted, but I don't have a DOI yet.

I've updated with Pypi version == 0.1.0 which corresponds to the changes from this review. There is also a paper-version branch for this version on the stella GitHub which has these changes. Here is a link to the [Zenodo page](https://zenodo.org/record/3970854#.XygVc5NKhAY) with the same title and authors as the JOSS paper. Let me know if there's anything I missed!!
