### 2020-08-03 #2347 @arfon
@openjournals/joss-eics - this paper is ready to publish once we have the AAS DOI.

/ cc @crawfordsm
