### 2020-08-04 #2347 @arfon
@afeinstein20 I just heard from AAS that your DOI will be `10.3847/1538-3881/abac0a`. Could you add this to your `paper.md`?
