### 2020-08-06 #2347 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
