### 2020-08-06 #2347 @afeinstein20
That was so fast! Thanks @arfon @pearsonkyle @astrobel @ekaterinailin for all of your help and paper/code comments!! You all made this such a fun and enjoyable experience 😄
